Marked/Tagged Content to Postgres DB (!)
Marked/tagged objects for users are now
stored in the database in a similar structure
to recent objects. Like recent objects, any
object can be tagged without necessitating
database changes, and it will immediately
replace the current marking functionality as
displayed in the Monkey Navigator.

This requires a new database table, which you
can re-export to pick up, or simply add it via
pgAdmin or something of the like using the
script below:

CREATE TABLE marked
(
  id integer NOT NULL,
  "type" character varying(255) NOT NULL,
  objectid integer NOT NULL,
  mark character varying(255) NOT NULL,
  userid integer NOT NULL,
  CONSTRAINT marked_pkey PRIMARY KEY (id),
  CONSTRAINT marked_userid_fkey FOREIGN KEY (userid)
      REFERENCES "user" (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
)

Change-Id: I39588a748eb73e1472cda7576f1525e3d83e197b
Reviewed-on: http://review.source.gogoego.com/1850
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>