Grid-based navigation header.
Instead of using the (deprecated) DataList, changed the
model to use Grids.  This allows for consistency and
a few layout tricks, including:

* Data Grouping
* Paging
* Double-click

First iteration here is missing a few key items, including

* Persistent selection upon bookmark load
* Marking
* Lock Status display for assessments

Which will re-appear in a later commit.

While it currently pages, Grids also allow us to think
about using other tricks to handle large amounts of
data, which we can revisit when functionality is a bit
more solid.  Using these tricks should not be a large
change once we are in a more solid state, as it may
only require updating the view a bit.

See #229
See #262
See #268

Change-Id: I9e499ee06268c0f26edaf63404ce4c6b872738d0
Reviewed-on: http://review.source.gogoego.com/1919
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>