Add edit information to AssessmentCache after save
After saving an assessment, field information is loaded
into the current assessment (IDs and such), but the edit
it's not an entire replacement, so some information (like
edits) get lost.

May need to revisit this later, but for now, explicitly
update the assessment edits for the current (cached)
assessment such that it appears in the Assessment Data
Browser after save.

Also, added a standard short and long format to the
FormattedDate API, which is used across all of SIS's
client interface to display dates.

See #286

Change-Id: I07a62b51aadbe8031b5ede159f25f140d0cb8d86
Reviewed-on: http://review.source.gogoego.com/1923
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>