Removed ubiquitous PanelManager + friends.
Was getting old really fast.  Removed the panel
manager and any classes that referenced it, which
didn't do much of anything anymore anyway.

Large commit, but much of this is simply the result
of organize imports.

Probably a few lines of useful code going away, but
it can be collected via version history if needed
for something in the future.  Most of this is either
moved to classes where it's currently used now, or
no longer needed.

Change-Id: Ice36b0580e92b368df5326f78eb77a123376d4cd
Reviewed-on: http://review.source.gogoego.com/1929
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>