Page URL support, code doc & cleanup
Added page URL support to allow for direct
navigation to sections within pages of more
complex interfaces, such as the Working Set
page.

Send the URL within the StateChangeEvent to
get this behavior.  Of course, it's up to
the implementing class to respond to the
given URL.

Specifically, this makes it possible to create
a new working set, then be taken directly to
the add taxa page.

Also, as a small improvement, make the taxon
name in the ADB page appear clickable
on mouse over.

See #273
See #267

Change-Id: I9ec77eb55ba3eb310bfd5c5451d46e41339b4a97
Reviewed-on: http://review.source.gogoego.com/1930
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>