Login Panel face lift, cleanup, fixes.
Fixed a bug that was forcing the login panel to
draw incorrectly when invalid credentials were
submitted or the user clicked the logout button.

Also, changed the login handler to use the API
to parse the profile of the logged-in user to
determine status as an SIS user.

Finally, removed the enable/disable login button
logic, replaced with a series of loading screens,
and did some code cleanup to remove many annoying
globally-defined variables, replacing them with
utility methods.  In addition, added utility
methods for clearing out the password field upon
unsuccessful login, and clearing everything upon
logout.

See #288

Change-Id: I86cb9082996fe1cadceac2594d3325490406e011
Reviewed-on: http://review.source.gogoego.com/1931
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>