Bug Fix: setState being called for setAssessment
This was an API change that was not properly
propogated.  setState used to take an assessment,
now takes a StateChangeEvent.  New setAssessment
object takes the assessment, so change all calls
to setState to use setAssessment.

This will fix a number of would-be NPEs, but will
particularly fix the taxon home page link from
the assessment data browser.

See #267

Change-Id: I7cac0c306cba9df364fa20186e851f005ea7655a
Reviewed-on: http://review.source.gogoego.com/1934
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>