Bug Fix: Zendesk Window authenticates correctly
Was sending a request to /zendesk/authn, which is our
local authentication, instead of /zendesk/login,
which does zendesk authentication.  Hence, the request
failed.  Unfortunately, in a surprising turn of events,
this failure was never propogated as an error message
to the user, so it seemed that "nothing happened."

Fixed the URL to go to the right place and pop an
error message if authentication fails for whatever
reason.  Also, added an icon for the bug reporting
tool to the gear menu.

See #263

Change-Id: I1c60f0c6db99ea37b2d6f6382a719fa6f33f13e6
Reviewed-on: http://review.source.gogoego.com/1935
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>