AssessmentFilter Bug Fix and Test Case
Fixed a bug related to assessment schema when
filtering asessments in working sets.  Also,
added a (poorly written, but correct) test
case that tests that this behavior is correct.

Once the test library is hooked up correctly to
the SIS software, we can change this test case
to use the Filter in a better way (directly).
For now, though, the case is accurate and works.

This change affect assessment iteration when
a working set is selected as well as what
appears in the navigation.

See #229
Test #266

Change-Id: Ib0737b483abc2b46ba70f0d781cf59738e0c94d7
Reviewed-on: http://review.source.gogoego.com/1943
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>