Search and Working Set Improvements
Working Set methods for adding taxa were broken and
had FIXMEs due to UI changes.  Made the appropriate
fixes here.

To do so also required updating search (for search to
add taxa method of working set).  Search already was
performing in underwhelming fashion and the code
was fairly inextensible.  Changed over from using
deprecated code to using Grid.  Interface remains
the same, nearly an unnoticeable difference in the
client, just faster with better support for checkboxes.

In doing this, had to refactor threats to make the
search there conform to the new API.

On the server, result nodes are served with the full
range of information needed to display in the grid,
which means taxa and assessments do not need to be
downloaded to the client proactively in order to
view the result list.

Finally, standardized the search mechanism to one
function in BodyContainer, and added a search cache
that can be accessed via the Gear menu.

See #198
See #234
See #271

Change-Id: I1f8c6e48c688deb41280647b2157d471c7e06c86
Reviewed-on: http://review.source.gogoego.com/1948
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>