Bug Fix: Existing PrimitiveFields not clearing.
There was a bug in the primitive field save model.
If no data was detected, instead of clearing/removing
the field, it was simply ignored.  This pattern works
when saving new fields, but for existing fields, it
means that if you cleared the field, it won't actually
make any changes, meaning the previous value is stuck.

Modified the code so that in these cases, the data is
removed, and the actual PrimitiveField model is removed
from the Field, when possible.  In the case of boolean
fields, this will cause the hasData check to fail, and
the field itself will be completely removed.

See #282

Change-Id: I0159f95dcdecc13e980c05ab6050050c36ee440b
Reviewed-on: http://review.source.gogoego.com/1955
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>