Sessions are now units of work, not global.
Session object was a global object, persisted
for the duration of the server life.  This is
wrong.  Sessions should be considered units
of work and be created and disposed of for
atomic operations.

Clearly, this is not easy.  Created utility
classes, TransactionRestlet and TransactionResource
that can be extended to simply make any actions
done within a particular handle() call a single
session transaction.  A new session is opened
each time.  If you aren't extending one of these
classes, you're on your own, but same rules apply.

Removed/refactored extra methods added in all DAO
classes such that a session object is required for
each method.

Code cleanup, as typical with these sorts of
refactorings, also happened.

Also changed in this commit is the VFS backup
storage; no longer does this take place. All
releveant data is stored in the database, only.
A replacement is in the works.

Change-Id: I8233c93b95d081f658c80d3bd95e084fd0445301
Reviewed-on: http://review.source.gogoego.com/1957
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>