Taxon images use thumbnails, dedicated ViewRestlet
Moved the lookup and parsing work done when retrieving
the taxon image for the taxon home page from the client
to the server.  The old means still remains in the API,
but a new means will allow the client to retrive the
primary image for a taxon without having to parse through
the XML to find the information and create the correct
URI for the image.  The URL

   {image-api-base}/images/view/thumb/{taxon-id}/primary

will yield a thumbnail, resized by any available GoGoEgo
resizer tool, of the primary image for the given taxon.
The thumbnail size can be specified by a size query
string parameter.

To get the full-sized primary image, simply hit:

   {image-api-base}/images/view/full/{taxon-id}/primary

If given a full path name instead of the "primary" keyword
at the end of the URL, that image will be  given.  The
thumbnailing and sizing options still apply.

Also, updated the client to use the new image view
restlet for image retrieval on the taxon home page.

See #244

Change-Id: I103061bd4b144be75c3a3fae05374bb5fec402cf
Reviewed-on: http://review.source.gogoego.com/1961
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>