ImageUtils port + Bug Fix for URI
Ported the ImageUtils class directly from
GoGoEgo in order to avoid having to copy all
images to the GoGoEgo VFS.  ImageUtils works
directly from the GoGoEgo VFS and it needs to
view the SIS VFS for these operations.

Also, fixed a bug in constructing the image
uri -- was pulling the remaining path, which
includes the query string.

Test #244

Change-Id: Ic32171a54093abecb2005433e5cac6d5cd3ebd1a
Reviewed-on: http://review.source.gogoego.com/1966
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>