Changed User Filter Popup to Window
The user filter popup is not cross-browser
compatible.  In at least Firefox, the text
fields seems to lose state when you click
them and try to type text.  This behavior
does not happen in (at least) Chrome.

Changed the popup div to a popup window in
order to defeat this odd behavior.

See #247

Change-Id: Ie8e4d22b1750f6d9fe49e129367a192a1f4571c5
Reviewed-on: http://review.source.gogoego.com/1972
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>