Removed paging from the navigation.
In place of navigation, lazily rendering of the
Grid (BufferView) is being used, which seems to
be pretty performant on respectably-sized data
sets.

A consequence of this was losing the free lunch
from GroupingView, so added custom grouping headers
group counts.  Added some styling, but more to
come in a later commit.  Initial styling is proof
of concept to display headers differently from
content.

See #296

Change-Id: I1d18425697d2667ef7eb58dd969440e196844cd5
Reviewed-on: http://review.source.gogoego.com/1988
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>