Taxon Images to DB, ImageViewer Fixes, Cleanup (!)
Taxon images are now related to the correct taxon
in the database, as opposed to a file-based
registry.  Images still live on the file system as
expected, but can be ported to another system without
much fuss.

While pushing this information to the database, also
updated the affected image manager client, fixing some
programmatic bugs and display issues regarding resizing
of thumbnails.  Ported the server to talk to the
database.

Finally, removed unused methods and classes.

You'll want the table below in your database:

CREATE TABLE taxonimage
(
  id integer NOT NULL,
  imagetype character varying(255) NOT NULL,
  identifier character varying(255) NOT NULL,
  primaryimage boolean NOT NULL,
  rating integer,
  weight double precision,
  caption text,
  credit character varying(1024),
  source character varying(1024),
  showredlist boolean,
  showsis boolean,
  taxonid integer NOT NULL,
  CONSTRAINT taxonimage_pkey PRIMARY KEY (id),
  CONSTRAINT taxonimage_taxonid_fkey FOREIGN KEY (taxonid)
      REFERENCES taxon (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
)

See #298
See #244

Change-Id: Ibe66c8c42e9d59bb5b19782ef0f63ea4285c1f6e
Reviewed-on: http://review.source.gogoego.com/1994
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>