Field model bug fixes for equals, setAssessment
Fixed two bugs with the Field model:

First, the equals method logic has changed -- the
generation ID is now only used when the Field's
id is set.  At this point, it's essentially the
same as the id.  Otherwise, a referential equality
measure is used to see if the two objects are the
same.  If so, we assume they are equal.  This should
be fine because two new (unsaved) Field objects
should be assumed unequal, but two Fields with the
same database-generated ID should always be equal.

Second, added a secondary method for setAssessment
that does NOT cascade to call the setAssessment
method for the child Fields.  This is used to ensure
that subfields are not associated with an assessment.
While I think this is the correct behavior all around,
the default setAssessment function still cascades.

Change-Id: I37a75a42376120011f117372ef0c42a7fcc4807f
Reviewed-on: http://review.source.gogoego.com/2029
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>