Use Loading Mask instead of Loading Screen in nav
Changed from using a loading screen to using a
simpler loading mask in the navigation areas.
This will still alert the user that data is
loading while allowing the user to continue to
click around.

This fixes the single-click, double-click issue
related to taxon when there are large amounts of
data and parsing being done when traversing
through working sets.

See #312

Change-Id: Ic272710f2f68d5b338ae5107051023a7ab73ec21
Reviewed-on: http://review.source.gogoego.com/2044
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>