Added detail data point to reduction basis fields.
Added a new field "detail" to the PopulationReduction
fields such that AOO, EOO, and Quality of Habitat can
be uniquely specifield when selection (c) is chosen.

The structures of these fields have been modified to
be related structures such that the detail field only
appear when the appropriate option is chosen.

See #319

Change-Id: Ia792c61f17fd2f7a692a369fc2dd4f5e57b2f6e7
Reviewed-on: http://review.source.gogoego.com/2050
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>