Set case sensitively when Postgres is lookups DB
The data conversion from H2 to Postgres creates tables
and columns in all uppercase, which makes many queries
fail in a case-sensitive environment.  An easy fix for
this is to set the DBSession case sensitivity to upper
when Postgres is indeed the session in use.

Perform a check for this and set the value appropriately.

Change-Id: I163e93edc8df43c30c025c29f0182c555867811f
Reviewed-on: http://review.source.gogoego.com/2052
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>