Client-side assessment save fixes.
Made the sync operation recursive such that child
fields get updated identifiers as well as top-level
ones.

TODO: The child fields are assigned "random" database
IDs because there's no way to map them from the client
to the server with a page refresh.  This is OK because
a subsequent save should correct all values, but it
would be nice to have consistent identifiers set between
client and server at some point down the line.
Change-Id: Iea5ea1995fe92484884712b0cb1982de41aa804a
Reviewed-on: http://review.source.gogoego.com/2059
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>