Field no longer required to have an assessment (!)
A field should have one parent; that parent
being either an assessment or another field, but
not both.  This is because forcing each field to
have an assessment as a parent causes replication
in the XML, since it will pull fields and subfields
as children of the assessment.  Makes for mighty
long XML when classification schemes are involved.

Hence, this commit changes the default behavior
of Field.setAssessment to NOT cascade; also, I've
updated the Field.hbm.xml file to set not-null to
FALSE.

This requires a database change.  You'll want to
remove the non-null restriction on the Field table's
assessment ID column.  You can do this via SQL, or,
if you have Postgres, simply right-click the column
to edit and uncheck the non-null box.

Change-Id: Ib89b38c7ddba115626c02593305703291f625249
Reviewed-on: http://review.source.gogoego.com/2084
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>