FloatPrimitiveField and SISNumber fixes
The FloatPrimitiveField was doing a bit of
damage to data integrity as its setRawValue
function was losing Float information for
large numbers by removing non-numeric
characters (like 'E').

Changed this so that FloatPrimitiveField's
setRawValue function only removes commas from
the value; this is OK since it already throws
a NumberFormatException, so any problem should
be handled by the caller.

Also, updated the SISNumber struture, since
it should now be handed a fully-qualified Float
value.  It instantiates a formatter based on
the data restrictions (0-1, year, none) and
uses that formatter on the value given when
populating the field.  True Floats are
displayed with up to four decimal places of
accuracy.

Change-Id: I6287be229858b47902f5d9174d282e1051701a44
Reviewed-on: http://review.source.gogoego.com/2113
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>