Bug Fix: Working Set Filter fail on Published Asms
Fixed a bug that was causing published assessments
to not be read in correct via the AssessmentFilterHelper
as allowed.  This was due to a logic error in that
a Working Set holds either draft OR published
assessments.  In fact, they can hold both at the same
time, in code.  In practice, this does not occur, but
support for it is around.

The bug assumed one or the other, and was failing in
the case of published assessments since some filters
seem to be coming over as both published and draft.
While it seems the bug is actually in the logic that
allows a client-created filter to be both draft and
published even when published is specifically
selected, the server code no longer falls over when
this occurs.

Change-Id: I88d1268149bb3ac159c2b9614ce2637673c82249
Reviewed-on: http://review.source.gogoego.com/2129
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>