Assessment Field Change Log (!)
Implemented a basic change tracking utility that
interrogates each save and logs the changes in
relation to the assessment, the field, and the
edit for that save.  It saves the field data in
it's entirety in model form.

In addition, added a server resouce with abilities
to show all edits for a given assessment, show all
changes for a given edit, and show all changes for
a given field (for a given assessment).

Finally, added a very simple viewer that shows
old versus new for a given edit in read-only
form.  It's raw and not pretty, but it works such
that we can see what exactly is "changing" each
time saves happen, and we have some reliance that
data should not get lost.

This requires a new table, which you can add via:

CREATE TABLE assessmentchange
(
  id integer NOT NULL,
  fieldname character varying(255) NOT NULL,
  "type" integer NOT NULL,
  editid integer NOT NULL,
  oldfieldid integer,
  newfieldid integer,
  assessmentid integer NOT NULL,
  CONSTRAINT assessmentchange_pkey PRIMARY KEY (id),
  CONSTRAINT assessmentchange_assessmentid_fkey FOREIGN KEY (assessmentid)
      REFERENCES assessment (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION,
  CONSTRAINT assessmentchange_editid_fkey FOREIGN KEY (editid)
      REFERENCES edit (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION,
  CONSTRAINT assessmentchange_newfieldid_fkey FOREIGN KEY (newfieldid)
      REFERENCES field (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION,
  CONSTRAINT assessmentchange_oldfieldid_fkey FOREIGN KEY (oldfieldid)
      REFERENCES field (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
);

Change-Id: I3f2b1ea53f856fb0d20e6bd877d2a513e333997e
Reviewed-on: http://review.source.gogoego.com/2141
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>