Locking table name change and fixes (!)
Changed the name of the lock table from "lock"
to "persistentlock" and changed a column name
from the (reserved word) "group" to "groupname".

Updated the PersistentLockRepository for these
changes, but changed the restlet to use the
Hibernate version directly now, as the problems
that were causing this to not work should be
fixed now.  If not, we can revert.

While this is a "change", technically, a new
table has been created, described below:

CREATE TABLE persistentlock
(
  id integer NOT NULL,
  lockid integer NOT NULL,
  date timestamp without time zone NOT NULL,
  "type" character varying(255) NOT NULL,
  groupname character varying(255) NOT NULL,
  userid integer NOT NULL,
  CONSTRAINT persistentlock_pkey PRIMARY KEY (id),
  CONSTRAINT fkba973042d2b1a896 FOREIGN KEY (userid)
      REFERENCES "user" (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
)

Change-Id: I3733b95f877b00fb9a4f5b85e44789a1ddcab6dd
Reviewed-on: http://review.source.gogoego.com/2157
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>