Partial and Full Assessment fetching.
The system fetches assessments a lot but
rarely does it ever need to pull the entire
data set for any given assessment.

Added a FetchMode to the AssessmentCache's
operations, allowing the caller to specify
how much data is actually needed.  A full
assessment is, predictably, the complete data
set, used with the Assessment Data Browser.
A partial assessment has a direct subset of
data as a full assessment, but is used everywhere
else assessment information is shown, such as
navigation, working set pages, and taxon pages.

This will allow for faster navigating and browsing
at the cost of having to download assessments
on-demand for editing purposes.

The server changes don't make much difference
here, but this change can be backed out of the
client at any time by just pulling FULL
assessments instead of partial ones, since it's
the same model.

Change-Id: I2328f1e33acf229875d85f5a438a67c2bd0e2f6c
Reviewed-on: http://review.source.gogoego.com/2163
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>