Additional class.scheme rendering improvements
Added a getPrefix function that can be used on
TreeDataRow to ensure the proper prefix is always
constructed.  It will only show prefixes for items
with row number that are actual numbers (which
excluded CountryOccurrence) that are greater than
zero.

Also, fixed an issue with the tree rendering where
a race condition could be had when attempting to
access the tree before it had been rendered to get
the selected rows.

See #322
See #255

Change-Id: I12c68e881242a3748e6cea35ff64e769326b6e8f
Reviewed-on: http://review.source.gogoego.com/2207
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>