Corrected Country Names
Corrected the country names in the lookups database
such that the parent-child relationship are now correct.

This correction is based off of the contents of the
CountryOccurrence.xml file, which is used to generate
the field on the client.  In theory, this is the way it
was done before, so there should be no missing information
or anything like that.

Included in this commit is the script used to parse that
XML and generate the appropriate SQL; while it was written
for CountryOccurrence, it should work on any Classification
Scheme, so if we need to regenerate any tables or data,
we can.

This issue is slightly related to the unresponsive script
error because since there were so many countries at the
root level (~300), it would not render correctly from the
database at the U/T level.  The normal field was not
affected because its data was read from file.

See #256

Change-Id: I1a26fe1f5a7395df76a6342dda3e20c1eaeaa52a
Reviewed-on: http://review.source.gogoego.com/2208
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>