Parameter support, additional debugging, iamges
Added support for form parameters directly in the
Converter object, such that these no longer need
to be individually parsed up-front; rather, that
can be taken care of by the implementing Converter.
E-mail support is also available, so results can be
e-mailed after a conversion completes (or doesn't).

Also, added/removed/cleaned debugging for the
AssessmentConverter.  Added more information where
needed and removed unnecessary lines.  Also, added
debug() and error() methods that specify what sort
of information is being printed, which also allows
easy filtering of information (such as turning off
all debugging for a run).

Finally, added a boolean flag to allow for images to
not be copied over via script.  Probably quicker to
simply to this via the command-line where available;
the option is still there to do so.

Change-Id: Idb29f03cbd7c78a06132e0ce3cbc31cd6a9dd4e2
Reviewed-on: http://review.source.gogoego.com/2227
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>