Lateral Move, StateManager supports forced refresh
For lateral move, and all taxonomic operations, added
support for a forced refresh that will ensure the taxon
home page gets updated if you are currently viewing a
page when making changes, even when making changes to
the current taxon.

Also, fixed a bug in Lateral Move that would make the page
get stuck if you hit move without selecting a taxon.

See #347

Change-Id: I8cc0293233a6a08bbd99dc206baac3cae804d5e6
Reviewed-on: http://review.source.gogoego.com/2233
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>