Code cleanup, correct ClassCastException, delete
Corrected an exception that caused this operation
to fail, and also did some code cleanup in anticipation
of having more work to do here (i.e. moving draft and
published assessments).

Also, addressed some issues with deleting taxa as well
as re-drawing the page when update is called and the
current taxon has not changed.

See #352

Change-Id: I7d2a5fc3cf5e1bcb6ca537f07732b5213e9a5527
Reviewed-on: http://review.source.gogoego.com/2239
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>