Add Profile via Manage Credits Window
Added client and server support for permitted
users to add a user profile.  Performs a first
name/last name search to check for existing
records, warning the user if applicable.  Then,
generates a faux user name and allows the user
to create the profile, and adds it to the list
of search results upon successful completion.

In doing this, and search functionality was added
to the ProfileSearchResource.  The profile search
perform an AND search in the preliminary query,
instead of a standard OR query, to eliminate names
like "John Doe" coming as a result when attempting
to add a "Joe Doe" profile.  The normal search
still works as normal (OR query).

Also, consolidated querying code between three
classes into a single utility class,
UserSearchController.  SearchResults model now
lives in here as well, along with the proper
search keywords.

Finally, for dependency reaons, moved the
AddUserWindow to the shared project, since the
AddProfileWindow depends on it.

See #44
See #67

Change-Id: I8c70fe2629e270eb14026bbe88e62487b5d6f0a8
Reviewed-on: http://review.source.gogoego.com/2257
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>