Persist ordering of credited users, allow re-order
Added an order subfield to each of the credited
users fields, and persist this information here in
CSV format after save such that the list can be
kept in designated order.

Also, added buttons to allow users to move items
up and down in each of the lists.

See #67
See #362

Change-Id: I7323a4d1e79cd6f3d5c60daeb75d3860daefe023
Reviewed-on: http://review.source.gogoego.com/2258
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>