Fixed bug with adding new taxa to working set.
This bug was order-sensitive, hence why it only
happened sometimes.  When downloading taxa from
the server, the "footprint" in the object layer
may be inconsistent; creating a new taxon from
a taxon with an inconsistent footprint causes the
child to also have this inconsistent footprint.
This means there may be only 3 entries in the
footprint when there should be 7.  Thus, the error
when attempting to sort.

Fixed this by using the non-object layer footprint
to construct new taxa.

This also fixes another issue with creating new
assessments for these inconsistent taxa.  Since
the taxa is length 4 and not 7, when checking to
ensure the footprint is correct length, it find
that it's not, and thinks the level is not a
species.

See #341
See #379

Change-Id: Ifa61ca08602b3ba9aa823638d9ec8e4eac9175d0
Reviewed-on: http://review.source.gogoego.com/2282
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>