Reference save behavior fixes, cleanup.
References at the global/assessment level now save appropriately.
Code before was not sending a request to the server, so nothing
was getting persisted.

References at the field level are also now getting saved in all
cases.  Before, saving a references to a field with no data was
not being properly persisted; the field would be deleted since
there was no data, and nothing would be saved.  Now, fields with
no data are allowed as long as they have a note or reference
attached.

Additionally, for the reference manager UI, ensure that changes
are shown in the search results after edits when appropriate, and
support ConflictExceptions when attempting to save a reference
that is used in one or more fields.  Related to his, a bug in
TransactionResouce was fixed that caused valid ResourceExceptions
to be swallowed and handled as uncaught exceptions.  This behavior
now replaces the "reference replace behavior."

Also, the "Show Field Names" button has been removed; instead,
it is always shown when viewing references at the assessment
level; otherwise, it is hidden.

This also fixes the "bug" with the number of references not being
updated; the number was correct, since references were not saving
correctly.

See #367
See #368
See #369
See #370

Change-Id: I174684a45654b04e66c5405efd9aa4a06cd0de52
Reviewed-on: http://review.source.gogoego.com/2295
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>