Add Profile gives tabular confirmation, API update
Added a table to the confirmation step to display the
users, which works better when there are many users
to display. (2)

Also, did a bit of code cleanup, including changing
the AddUserWindow API to accept a ComplexListener instead
of a GenericCallback -- there is nothing to do in the
onFailure method (uncalled) here; don't grab a hammer when
you don't have a nail.

See #335

Change-Id: I32eda643d7dadb000d4402c197cf21ad4aa5b077
Reviewed-on: http://review.source.gogoego.com/2296
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>