Bug Fix: Remove taxon throwing silent NPE
When removing a taxon, the Taxon Home Page would try to
redraw the parent, but instead of calling the in-memory
representation to get the parent, it was checking the
cache, which no longer has the node, so null is returned.
Unfortunately, the exception thrown is silent, so other
operations that would have happened do not, and other bad
things happen down the line.

This should fix that bug, and any other bugs related to
badness happening after a remove operation.

See #384

Change-Id: I0e747a19e30baa92948ec9d957cbda3360f53098
Reviewed-on: http://review.source.gogoego.com/2304
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>