Bug Fix: Newly added credits not re-orderable.
If an explicit order was already created, a newly
dropped user was not inserted into the ordering
scheme, and a subsequent explicit ordering of that
user would cause an IndexOutOfBoundsException, since
the user was not in the ordering scheme.

Changed logic to destroy and rebuild that ordering
scheme whenever a new user or users is dropped into
a drop target.

See #362

Change-Id: Icaf07efbc63f1a6ac123804121c6d5339d736ac2
Reviewed-on: http://review.source.gogoego.com/2308
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>