Ensure page iteration is consistent with nav.
Sorting methods used in the navigation is now
consistent with the the methods used when iterating
through items once you land on a data page.  So,
when using previous and next buttons on working set,
taxon, and assessment pages, the items will no longer
"skip," rather, they will go in the same order as they
appear in the nav.

This method can and should be optimized in the future,
but will run performance tests with this inital version
to see where, if any, the bottlenecks are.

See #308

Change-Id: I94ed1fef2d0659074ee23f509e2ce28765e41c3c
Reviewed-on: http://review.source.gogoego.com/2312
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>