Ensure relationships created for Field.deepCopy.
Parental relationships for fields, notes, references,
and primitive fields were not being properly generated
for Field.deepCopy method, causing some calls to fail
with database errors.  This was not an issue with client-
server calls since the relationships are properly
constructed with a fromXML call, but using this method
on the server explicitly caused bad results.

Specifically for #358, it was causing batch assessments
to not be able to be generated from templates, because
it would fail to save the new assessments to the database
since relationships were being set up incorrectly.

See #358

Change-Id: I8d4f506f5feb5a4a0514e737315eab8ecd394b74
Reviewed-on: http://review.source.gogoego.com/2315
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>