ReferenceCache now global, code cleanup.
The ReferenceCache now caches references
irrespective to their relationship to an assessment,
since references can appear on taxa, synonyms, and
more.

The client caches an assessment anytime it is parsed
via the ClientReferenceParser (essentially, when
Reference.fromXML is called client-side), and that
representation is displayed in the bibliography.
Ths ensures a consistent view for references when
looking at a reference used in multiple places.

Also, reference submission resource now ensure that
only reference properties can be updated for existing
refereces, not data relationships.

See #386

Change-Id: I15c2d9c06d34bf3a5ada89a6b8cab7bc2342b333
Reviewed-on: http://review.source.gogoego.com/2317
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>