Nav: Catch exception when parsing malformed data
For the taxon listing, errors can be thrown with a bad
taxon footprint that breaks the nav for the entire user
session.  Added a handle to catch this exception, but
the cause of this should be tracked down.  Want this for
now in order to allow the system to continue working.

Change-Id: I1de8a4630157579256baac3b2baf52e27abb25e4
Reviewed-on: http://review.source.gogoego.com/2323
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>