Keep current page selected when navigating DEM.
Current page was getting lost when moving from one
assessment to another; furthermore, the current page
was selected, so you could not re-select it without
going to a new page, the going back.  Implemented the
lingering TODO to fix this, and also ensure that the
page DOES reset when moving between assessments that
are in different schemas.

Change-Id: Ic9098aee948fd5d4e3ea8612cb396a03f6424367
Reviewed-on: http://review.source.gogoego.com/2325
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>